split helper methods
